+++ src/transformers/models/auto/modeling_auto.py
@@ -847,7 +847,9 @@ _MODEL_FOR_ZERO_SHOT_IMAGE_CLASSIFICATION_MAPPING_NAMES = OrderedDict(
 MODEL_FOR_BACKBONE_MAPPING_NAMES = OrderedDict(
     [
         # Backbone mapping
+        ("maskformer-swin", "MaskFormerSwinBackbone"),
         ("resnet", "ResNetBackbone"),
+        ("swin", "MaskFormerSwinBackbone"),
     ]
 )
