+++ src/transformers/models/maskformer/__init__.py
@@ -20,7 +20,10 @@ from typing import TYPE_CHECKING
 from ...utils import OptionalDependencyNotAvailable, _LazyModule, is_torch_available, is_vision_available
 
 
-_import_structure = {"configuration_maskformer": ["MASKFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP", "MaskFormerConfig"]}
+_import_structure = {
+    "configuration_maskformer": ["MASKFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP", "MaskFormerConfig"],
+    "configuration_maskformer_swin": ["MASKFORMER_SWIN_PRETRAINED_CONFIG_ARCHIVE_MAP", "MaskFormerSwinConfig"],
+}
 
 try:
     if not is_vision_available():
@@ -43,9 +46,15 @@ else:
         "MaskFormerModel",
         "MaskFormerPreTrainedModel",
     ]
+    _import_structure["modeling_maskformer_swin"] = [
+        "MaskFormerSwinBackbone",
+        "MaskFormerSwinPreTrainedModel",
+        "MaskFormerSwinModel",
+    ]
 
 if TYPE_CHECKING:
     from .configuration_maskformer import MASKFORMER_PRETRAINED_CONFIG_ARCHIVE_MAP, MaskFormerConfig
+    from .configuration_maskformer_swin import MaskFormerSwinConfig
 
     try:
         if not is_vision_available():
@@ -65,6 +74,11 @@ if TYPE_CHECKING:
             MaskFormerForInstanceSegmentation,
             MaskFormerModel,
             MaskFormerPreTrainedModel,
+        )
+        from .modeling_maskformer_swin import (
+            MaskFormerSwinBackbone,
+            MaskFormerSwinModel,
+            MaskFormerSwinPreTrainedModel,
         )
